refactored container, removing PSR and adding interface
Refs #15928
